Extract baseload calculation code from AMRData
